requirements.txt
appdirs==1.4.4
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
clingo==5.6.2
contourpy==1.1.1
cycler==0.12.1
dgl==1.1.2
docker-pycreds==0.4.0
fastjsonschema==2.19.0
filelock==3.13.0
fonttools==4.43.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
graphviz==0.20.1
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
joblib==1.3.2
jsonpickle==3.0.2
jsonschema==4.20.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
mpmath==1.3.0
nbformat==5.9.2
networkx==3.2.1
numpy==1.26.0
pandas==2.1.2
Pillow==10.1.0
plotly==5.18.0
protobuf==4.25.0
pyaml==23.9.7
pycparser==2.21
PyILP==0.0.1
pyparsing==3.1.1
pyvis==0.3.2
PyYAML==6.0.1
referencing==0.31.1
requests==2.31.0
rpds-py==0.13.2
scikit-learn==1.3.2
scipy==1.11.3
sentry-sdk==1.34.0
setproctitle==1.3.3
smmap==5.0.1
sympy==1.12
tenacity==8.2.3
texttable==1.7.0
threadpoolctl==3.2.0
torch==2.1.1
torch_geometric==2.4.0
torchview==0.2.6
torchvision==0.16.1
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.7
wandb==0.16.0
zipp==3.17.0